Add files via upload
This is a text based adventure game we are working on in order to navigate through rooms, drop/get items, and move through locked doors. Eventually we will be manipulating containers, ensuring the appropriate containers cannot be seen through if closed, and the respective see through containers are just that. This is a work in progress and will be updated regularly in the coming weeks.